feat: implement login service and enhance user deletion error handling

+++ services/user_services.py
@@ -2,8 +2,11 @@ from models.user import User
 from motor.motor_asyncio import AsyncIOMotorDatabase 
 from schemas.user import UserCreate, UserDelete, DeleteResponse
 from core.security import get_password_hash
-from typing import Optional, Any
-from exceptions.exceptions import UserAlreadyExistsError, NicknameAlreadyExistsError, UserNotFoundError
+from typing import Optional
+from schemas.token import Token
+from fastapi.security import OAuth2PasswordRequestForm
+from core.security import verify_password, create_access_token
+from exceptions.exceptions import UserAlreadyExistsError, NicknameAlreadyExistsError, UserNotFoundError,UnauthorizedError,IncorrectPasswordError
 
 async def create_user(user_data: UserCreate, db: AsyncIOMotorDatabase) -> User:
     existing_user = await get_user_by_email(user_data.email, db)
@@ -21,15 +24,28 @@ async def create_user(user_data: UserCreate, db: AsyncIOMotorDatabase) -> User:
     await db["users"].insert_one(user_dict)
     return User(**user_dict)
 
-async def delete_user(user: UserDelete, db: AsyncIOMotorDatabase) ->  DeleteResponse:
+async def login_service(form_data: OAuth2PasswordRequestForm, db: AsyncIOMotorDatabase) -> Token:
+    user = await get_user_by_nickname(form_data.username, db)
+    if not user:
+        raise UserNotFoundError(email=form_data.username)
+
+    if not verify_password(form_data.password, user.hashed_password):
+        raise IncorrectPasswordError(nickname=form_data.username)
+
+    access_token = create_access_token({"sub": user.email})
+    return Token(access_token=access_token, token_type="bearer")
+
+async def delete_user(user: UserDelete, db: AsyncIOMotorDatabase, current_user: User) -> DeleteResponse:
     existing_user = await get_user_by_email(user.email, db)
     if not existing_user:
         raise UserNotFoundError(email=user.email)
-    
+
+    if current_user.email != user.email and not getattr(current_user, "is_admin", False):
+        raise UnauthorizedError(user=current_user.email)
+
     await db["users"].delete_one({"email": user.email})
     return DeleteResponse(success = True, detail=f"User {user.email} deleted successfully")
         
-
 async def get_user_by_email(email: str, db: AsyncIOMotorDatabase) -> Optional[User]:
     user_data = await db["users"].find_one({"email": email})
     if user_data:
